+++ lib/simplepy.dart
@@ -3,5 +3,6 @@ library;
 export 'src/parser.dart' show Parser, ParseError;
 export 'src/interpreter.dart'
     show Interpreter, PyList, PyTuple, PyFunction, PyInstance, NativeFunction;
+export 'src/pynum.dart' show PyNum;
 export 'src/lexer.dart' show Lexer, LexerError;
 export 'src/ast_nodes.dart' show Stmt;